requirements.txt
flask
ibm-generative-ai
python-dotenv
